Notifications - enhance tests
